// frontend/src/lib/services/backend-api-generated/models/category.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bakebee Admin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { SubCategory } from './sub-category';

export interface Category {
    'id'?: string;
    'category_name'?: string;
    'is_active'?: boolean;
    'sub_categories'?: Array<SubCategory>;
}